# pkg/unpackerr/openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "Unpackerr API",
    "version": "1.0.0",
    "description": "HTTP API for Unpackerr. Login returns an admin API key and a session cookie. After that, send X-Api-Key (or Authorization: Bearer). This spec contains no secret values."
  },
  "servers": [
    {
      "url": "/",
      "description": "This Unpackerr instance; url is rewritten to the configured urlbase when the spec is served"
    }
  ],
...
  "security": [
    {"apiKey": []},
    {"bearer": []},
    {"session": []}
  ],
  "components": {
    "securitySchemes": {
      "apiKey": {
        "type": "apiKey",
        "in": "header",
        "name": "X-Api-Key"
      },
      "bearer": {
        "type": "http",
        "scheme": "bearer"
      },
      "session": {
        "type": "apiKey",
        "in": "cookie",
        "name": "session"
      }
    },
    "schemas": {
      "Error": {
        "type": "object",
        "properties": {
          "error": {"type": "string"},
          "auth": {"type": "string", "description": "UI auth mode on 401 when password login is disabled: header or noauth"}
        }
      },
...
      "Stats": {
        "type": "object",
        "properties": {
          "waiting": {"type": "integer"},
          "queued": {"type": "integer"},
          "extracting": {"type": "integer"},
          "failed": {"type": "integer"},
          "extracted": {"type": "integer"},
          "imported": {"type": "integer"},
          "deleted": {"type": "integer"},
          "hookOK": {"type": "integer"},
          "hookFail": {"type": "integer"},
          "cmdOK": {"type": "integer"},
          "cmdFail": {"type": "integer"},
          "retries": {"type": "integer"},
          "finished": {"type": "integer"},
          "starrs": {"type": "integer", "description": "Configured Sonarr+Radarr+Lidarr+Readarr instance count"},
          "folders": {"type": "integer", "description": "Configured watch-folder count"},
          "webhooks": {"type": "integer", "description": "Configured webhook count"},
          "cmdhooks": {"type": "integer", "description": "Configured command-hook count"},
          "stackFS": {"$ref": "#/components/schemas/BufferStat", "description": "Folder-watch filesystem event channel"},
          "stackXtractr": {"$ref": "#/components/schemas/BufferStat", "description": "Starr extract callback channel"},
          "stackFolder": {"$ref": "#/components/schemas/BufferStat", "description": "Folder-watch extract callback channel"},
          "stackHook": {"$ref": "#/components/schemas/BufferStat", "description": "Webhook and command-hook delivery channel"},
          "stackDel": {"$ref": "#/components/schemas/BufferStat", "description": "File-delete job channel"},
          "stackTask": {"$ref": "#/components/schemas/BufferStat", "description": "Config apply and queue-action channel"},
          "active": {"type": "integer", "description": "Compatibility count of non-terminal queue items"},
          "completed": {"type": "integer", "description": "Compatibility count of retained completed history rows"},
          "webhookOK": {"type": "integer", "description": "Compatibility alias for hookOK"},
          "webhookFailed": {"type": "integer", "description": "Compatibility alias for hookFail"},
          "cmdhookOK": {"type": "integer", "description": "Compatibility alias for cmdOK"},
          "cmdhookFailed": {"type": "integer", "description": "Compatibility alias for cmdFail"},
          "uptime": {"type": "string", "description": "Compatibility uptime display"},
          "generatedAt": {"type": "string", "format": "date-time", "description": "Compatibility response generation time"},
          "starrQueues": {
            "type": "array",
            "description": "Last activity-queue poll per configured Starr app",
            "items": {
              "type": "object",
              "properties": {
                "app": {"type": "string"},
                "name": {"type": "string"},
                "url": {"type": "string"},
                "queued": {"type": "integer", "description": "Starr TotalRecords"},
                "retrieved": {"type": "integer", "description": "Records returned this poll"},
                "complete": {"type": "integer", "description": "Retrieved items with Status completed"},
                "match": {"type": "integer", "description": "Completed items whose protocol is in this instance's protocol list"},
                "issues": {"type": "integer", "description": "Retrieved items that failed, warned, or have a tracked-download error/warning"},
                "downloading": {"type": "integer", "description": "Retrieved items with Status downloading"},
                "updatedAt": {"type": "string", "format": "date-time", "description": "When this instance's queue snapshot was last fetched"},
                "error": {"type": "string"}
              }
            }
          }
        }
      },
...
      "SystemInfo": {
        "type": "object",
        "properties": {
          "version": {"type": "string"},
          "revision": {"type": "string"},
          "started": {"type": "string", "format": "date-time"},
          "uptime": {"type": "string"},
          "listenAddr": {"type": "string"},
          "urlbase": {"type": "string"},
          "auth": {"type": "string"},
          "metrics": {"type": "boolean"},
          "configFile": {"type": "string"},
          "hostname": {"type": "string"},
          "goos": {"type": "string"},
          "logs": {"type": "string", "description": "Unique parent folders of the app and HTTP log files, comma-separated when they differ"}
        }
      },
      "FieldHelp": {
        "type": "object",
        "properties": {
          "short": {"type": "string"},
          "uishort": {"type": "string"},
          "desc": {"type": "string"},
          "env": {"type": "string"}
        }
      },
      "LiveExport": {
        "type": "object",
        "properties": {
          "text": {"type": "string"}
        }
      },
...
      "IDRequest": {
        "type": "object",
        "required": ["id"],
        "properties": {
          "id": {"type": "string", "description": "Queue or history item id (usually a filesystem path)"}
        }
      },
      "StatusReply": {
        "type": "object",
        "properties": {
          "status": {"type": "string"},
          "id": {"type": "string"}
        }
      },
...
      "ConfigWriteReply": {
        "type": "object",
        "properties": {
          "status": {"type": "string"},
          "restartRequired": {"type": "boolean"}
        }
      },
...
      "BrowseDir": {
        "type": "object",
        "properties": {
          "sep": {"type": "string", "description": "filepath.Separator on this host"},
          "path": {"type": "string", "description": "Directory being listed; empty on Windows volume-root listing"},
          "mom": {"type": "string", "description": "Parent directory; empty at a volume root"},
          "dirs": {"type": "array", "items": {"type": "string"}},
          "files": {"type": "array", "items": {"type": "string"}},
          "error": {"type": "string", "description": "Set when the requested path was unreadable but the parent listed successfully"}
        }
      },
      "BrowseCreate": {
        "type": "object",
        "required": ["path"],
        "properties": {
          "path": {"type": "string", "description": "Folder to create (MkdirAll 0755); existing folders succeed"}
        }
      },
      "BrowseTemplate": {
        "type": "object",
        "required": ["path", "template"],
        "properties": {
          "path": {"type": "string", "description": "Destination file; basename must be unpackerr-webhook-<template>.tmpl"},
          "template": {"type": "string", "enum": ["notifiarr", "discord", "telegram", "slack", "pushover", "gotify"]}
        }
      },
      "BrowseTemplateResult": {
        "type": "object",
        "properties": {
          "path": {"type": "string", "description": "Absolute path written"}
        }
      }
    },
    "responses": {
      "Unauthorized": {
        "description": "Missing or invalid credentials",
        "content": {"application/json": {"schema": {"$ref": "#/components/schemas/Error"}}}
      },
      "Forbidden": {
        "description": "Authenticated but missing permission",
        "content": {"application/json": {"schema": {"$ref": "#/components/schemas/Error"}}}
      },
      "BadRequest": {
        "description": "Invalid JSON or validation error",
        "content": {"application/json": {"schema": {"$ref": "#/components/schemas/Error"}}}
      },
      "Conflict": {
        "description": "The file already exists",
        "content": {"application/json": {"schema": {"$ref": "#/components/schemas/Error"}}}
      },
      "GatewayTimeout": {
        "description": "Request canceled or timed out while waiting on the main loop",
        "content": {"application/json": {"schema": {"$ref": "#/components/schemas/Error"}}}
      },
      "PersistError": {
        "description": "Failed to write the change to disk; it will not survive a restart",
        "content": {"application/json": {"schema": {"$ref": "#/components/schemas/Error"}}}
      },
...
    }
  },
  "paths": {
...
    "/api/stats": {
      "get": {
        "tags": ["system"],
        "summary": "Queue counts for Homepage and the dashboard",
        "description": "Requires system:stats:read",
        "responses": {
          "200": {
            "description": "Counts",
            "content": {"application/json": {"schema": {"$ref": "#/components/schemas/Stats"}}}
          },
          "401": {"$ref": "#/components/responses/Unauthorized"},
          "403": {"$ref": "#/components/responses/Forbidden"}
        }
      }
    },
    "/api/system": {
      "get": {
        "tags": ["system"],
        "summary": "Version, uptime, listen address, auth mode, config file path",
        "description": "Requires system:info:read",
        "responses": {
          "200": {
            "description": "System info",
            "content": {"application/json": {"schema": {"$ref": "#/components/schemas/SystemInfo"}}}
          },
          "401": {"$ref": "#/components/responses/Unauthorized"},
          "403": {"$ref": "#/components/responses/Forbidden"}
        }
      }
    },
    "/api/system/export": {
      "get": {
        "tags": ["system"],
        "summary": "Live config rundown (startup-log style, secrets omitted)",
        "description": "Requires system:info:read. Same shape of information as the startup log: apps, folders, intervals, webserver. Each section also needs config:{section}:read (or *); otherwise the dump prints an omit line naming the missing permission. API keys are present/absent booleans, UI password is the auth type only.",
        "responses": {
          "200": {
            "description": "Plain-text rundown in JSON",
            "content": {"application/json": {"schema": {"$ref": "#/components/schemas/LiveExport"}}}
          },
          "504": {"$ref": "#/components/responses/GatewayTimeout"},
          "401": {"$ref": "#/components/responses/Unauthorized"},
          "403": {"$ref": "#/components/responses/Forbidden"}
        }
      }
    },
    "/api/browse": {
      "get": {
        "tags": ["files"],
        "summary": "List files and folders",
        "description": "Requires system:browse:read. Empty dir expands to the process home directory; on Windows empty, / or \\ lists mounted C:–Z: drives that exist. A file path lists its parent. An unreadable path that has a readable parent returns 200 with error set. Both unreadable is 406. mom is empty at a volume root.",
        "parameters": [
          {
            "name": "dir",
            "in": "query",
            "required": false,
            "schema": {"type": "string"},
            "description": "Directory or file path; ~ is expanded"
          }
        ],
        "responses": {
          "200": {
            "description": "Directory contents",
            "content": {"application/json": {"schema": {"$ref": "#/components/schemas/BrowseDir"}}}
          },
          "406": {"description": "Path and parent are unreadable", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/Error"}}}},
          "401": {"$ref": "#/components/responses/Unauthorized"},
          "403": {"$ref": "#/components/responses/Forbidden"}
        }
      },
      "post": {
        "tags": ["files"],
        "summary": "Create a folder",
        "description": "Requires system:browse:write. Uses MkdirAll 0755. Existing folders succeed. Returns the listing of the new (or existing) folder.",
        "requestBody": {
          "required": true,
          "content": {"application/json": {"schema": {"$ref": "#/components/schemas/BrowseCreate"}}}
        },
        "responses": {
          "200": {
            "description": "Directory contents after create",
            "content": {"application/json": {"schema": {"$ref": "#/components/schemas/BrowseDir"}}}
          },
          "400": {"$ref": "#/components/responses/BadRequest"},
          "500": {"description": "Create failed", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/Error"}}}},
          "401": {"$ref": "#/components/responses/Unauthorized"},
          "403": {"$ref": "#/components/responses/Forbidden"}
        }
      }
    },
    "/api/browse/template": {
      "post": {
        "tags": ["files"],
        "summary": "Write a built-in webhook template file",
        "description": "Requires system:browse:write. Writes unpackerr-webhook-<template>.tmpl with the built-in Go template. Fails if the file already exists.",
        "requestBody": {
          "required": true,
          "content": {"application/json": {"schema": {"$ref": "#/components/schemas/BrowseTemplate"}}}
        },
        "responses": {
          "201": {
            "description": "File written",
            "content": {"application/json": {"schema": {"$ref": "#/components/schemas/BrowseTemplateResult"}}}
          },
          "400": {"$ref": "#/components/responses/BadRequest"},
          "409": {"$ref": "#/components/responses/Conflict"},
          "500": {"description": "Write failed", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/Error"}}}},
          "401": {"$ref": "#/components/responses/Unauthorized"},
          "403": {"$ref": "#/components/responses/Forbidden"}
        }
      }
    },
...
    "/api/queue": {
      "get": {
        "tags": ["queue"],
        "summary": "In-flight extracts",
        "description": "Requires system:queue:read",
        "responses": {
          "200": {
            "description": "Live queue",
            "content": {"application/json": {"schema": {"type": "array", "items": {"$ref": "#/components/schemas/QueueItem"}}}}
          },
          "401": {"$ref": "#/components/responses/Unauthorized"},
          "403": {"$ref": "#/components/responses/Forbidden"}
        }
      }
    },
    "/api/queue/item": {
      "get": {
        "tags": ["queue"],
        "summary": "One in-flight extract, including archive and extracted file lists",
        "description": "Requires system:queue:read. File lists are omitted from GET /api/queue and progress frames so large extracts stay cheap on the wire. Id is a query parameter because it is usually a filesystem path.",
        "parameters": [
          {
            "name": "id",
            "in": "query",
            "required": true,
            "schema": {"type": "string"},
            "description": "Queue item id (usually a filesystem path)"
          }
        ],
        "responses": {
          "200": {
            "description": "Queue item with file lists",
            "content": {"application/json": {"schema": {"$ref": "#/components/schemas/QueueItem"}}}
          },
          "400": {"$ref": "#/components/responses/BadRequest"},
          "404": {"description": "Not in the live map", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/Error"}}}},
          "401": {"$ref": "#/components/responses/Unauthorized"},
          "403": {"$ref": "#/components/responses/Forbidden"}
        }
      }
    },
    "/api/queue/retry": {
      "post": {
        "tags": ["queue"],
        "summary": "Retry an extractfailed item",
        "description": "Requires system:queue:write. Id is a path in the JSON body.",
        "requestBody": {
          "required": true,
          "content": {"application/json": {"schema": {"$ref": "#/components/schemas/IDRequest"}}}
        },
        "responses": {
          "200": {
            "description": "Queued for retry",
            "content": {"application/json": {"schema": {"$ref": "#/components/schemas/StatusReply"}}}
          },
          "404": {"description": "Not in the live map", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/Error"}}}},
          "409": {"description": "Not extractfailed", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/Error"}}}},
          "400": {"$ref": "#/components/responses/BadRequest"},
          "504": {"$ref": "#/components/responses/GatewayTimeout"},
          "401": {"$ref": "#/components/responses/Unauthorized"},
          "403": {"$ref": "#/components/responses/Forbidden"}
        }
      }
    },
    "/api/queue/forget": {
      "post": {
        "tags": ["queue"],
        "summary": "Drop a terminal queue item from tracking",
        "description": "Requires system:queue:write. Only extractfailed, extractednothing, imported, deleted, and deletefailed items can be forgotten.",
        "requestBody": {
          "required": true,
          "content": {"application/json": {"schema": {"$ref": "#/components/schemas/IDRequest"}}}
        },
        "responses": {
          "200": {
            "description": "Forgotten",
            "content": {"application/json": {"schema": {"$ref": "#/components/schemas/StatusReply"}}}
          },
          "404": {"description": "Not in the live map", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/Error"}}}},
          "409": {"description": "Still in progress", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/Error"}}}},
          "400": {"$ref": "#/components/responses/BadRequest"},
          "504": {"$ref": "#/components/responses/GatewayTimeout"},
          "401": {"$ref": "#/components/responses/Unauthorized"},
          "403": {"$ref": "#/components/responses/Forbidden"}
        }
      }
    },
    "/api/history": {
      "get": {
        "tags": ["history"],
        "summary": "Completed and failed items (newest first)",
        "description": "Requires system:history:read. Completed and failed rows are capped by keep_history. In-flight checkpoints are not listed and do not count against that cap.",
        "responses": {
          "200": {
            "description": "History",
            "content": {"application/json": {"schema": {"type": "array", "items": {"$ref": "#/components/schemas/HistoryRecord"}}}}
          },
          "401": {"$ref": "#/components/responses/Unauthorized"},
          "403": {"$ref": "#/components/responses/Forbidden"}
        }
      }
    },
    "/api/history/delete": {
      "post": {
        "tags": ["history"],
        "summary": "Delete one finished history row",
        "description": "Requires system:history:write. In-progress extracts (queued, extracting, extracted, deleting) are resume checkpoints and cannot be deleted.",
        "requestBody": {
          "required": true,
          "content": {"application/json": {"schema": {"$ref": "#/components/schemas/IDRequest"}}}
        },
        "responses": {
          "200": {
            "description": "Deleted",
            "content": {"application/json": {"schema": {"$ref": "#/components/schemas/StatusReply"}}}
          },
          "404": {"description": "Unknown id", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/Error"}}}},
          "409": {"description": "Id is still in progress", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/Error"}}}},
          "400": {"$ref": "#/components/responses/BadRequest"},
          "500": {"$ref": "#/components/responses/PersistError"},
          "401": {"$ref": "#/components/responses/Unauthorized"},
          "403": {"$ref": "#/components/responses/Forbidden"}
        }
      }
    },
    "/api/history/clear": {
      "post": {
        "tags": ["history"],
        "summary": "Clear finished history rows",
        "description": "Requires system:history:write. Does not drop in-progress extracts kept for restart resume.",
        "responses": {
          "200": {
            "description": "Cleared",
            "content": {"application/json": {"schema": {"$ref": "#/components/schemas/StatusReply"}}}
          },
          "500": {"$ref": "#/components/responses/PersistError"},
          "401": {"$ref": "#/components/responses/Unauthorized"},
          "403": {"$ref": "#/components/responses/Forbidden"}
        }
      }
    },
    "/api/config/env": {
      "get": {
        "tags": ["config"],
        "summary": "Environment variables that overlaid config at startup",
        "description": "Requires authentication. Keys are UN_* suffixes with the same case as the TOML map key (DEBUG, SONARR_0_URL, SONARR_uhd_URL, WEBSERVER_ROLES_stats_PERMISSIONS_0). Values are the env strings cnfg wrote into fields. Password, API key, webhook token, and webserver api key values are blank unless the caller has *. WEBSERVER_UI_PASSWORD is always blank (the key stays so the UI can lock the field).",
        "responses": {
          "200": {
            "description": "Map of env suffix to value",
            "content": {"application/json": {"schema": {
              "type": "object",
              "additionalProperties": {"type": "string"}
            }}}
          },
          "401": {"$ref": "#/components/responses/Unauthorized"}
        }
      }
    },
    "/api/config/help": {
      "get": {
        "tags": ["config"],
        "summary": "Field help from definitions.yml",
        "description": "Requires authentication. English short/desc/env for each config field, keyed as config.<section>.<name>. The SPA uses this for ? tooltips; locale files may override.",
        "responses": {
          "200": {
            "description": "Map of field id to help",
            "content": {"application/json": {"schema": {
              "type": "object",
              "additionalProperties": {"$ref": "#/components/schemas/FieldHelp"}
            }}}
          },
          "401": {"$ref": "#/components/responses/Unauthorized"}
        }
      }
    },
    "/api/config/{section}": {
      "parameters": [
        {
          "name": "section",
          "in": "path",
          "required": true,
          "schema": {
            "type": "string",
            "enum": ["general", "webserver", "sonarr", "radarr", "lidarr", "readarr", "folders", "hooks", "webhooks", "cmdhooks"]
          }
        }
      ],
      "get": {
        "tags": ["config"],
        "summary": "Read the on-disk config section",
        "description": "Requires config:{section}:read. Returns file-shaped values (filepath: prefixes kept). Starr, folders.folder, webhooks, and cmdhooks are objects keyed by slug (v1.0.0, September 2026); env-only slugs are omitted. Starr API keys are visible as stored. Unpackerr webserver apiKeys[].key is returned only to callers with *; PUT keeps a blank key of the same name. ui_password is !!cryptd!!, webauth, noauth, or filepath:; plaintext user:pass is blanked. PUT a new password as user:<kdf-hex> (same digest as login), not plaintext. Use GET /api/config/{section}/live for expanded running values. PUT this payload back to save.",
        "responses": {
          "200": {
            "description": "On-disk section payload",
            "content": {"application/json": {"schema": {"$ref": "#/components/schemas/ConfigSection"}}}
          },
          "404": {"description": "Unknown section", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/Error"}}}},
          "401": {"$ref": "#/components/responses/Unauthorized"},
          "403": {"$ref": "#/components/responses/Forbidden"}
        }
      },
      "put": {
        "tags": ["config"],
        "summary": "Replace one config section and rewrite the TOML file",
        "description": "Requires config:{section}:write. Body is the file-shaped GET payload (not /live). Starr/hooks are slug maps; {} clears file instances (legacy [] still loads as keys 0, 1, …). Folders is {buffer, folder}; empty {} is 400, folder may be {}. PUT writes the request body as the file document; live is clone(file)+ParseENV so env-only slugs still appear after {}. Unknown JSON fields, extra values, empty general/webserver objects, and nil list entries return 400. Validation, the TOML write, and the live apply run on the main loop; a persist failure is 500 and leaves runtime unchanged. A filepath: string is expanded for the running process and kept as filepath: on disk. Adding or changing a filepath: is allowed; a missing secret file is 400. Replacing filepath: with a literal value is allowed. Empty apiKeys[].key keeps the existing key of the same name so a redacted GET can round-trip without *. webserver uiPassword on PUT is !!cryptd!!, webauth, noauth, filepath:, or user:<64-char PBKDF2 hex> (the same kdf as login); plaintext user:pass is 400. Changing the live password hash or auth type while local password auth is on requires uiCurrentKdf (KDF of the current username+password). General interval changes reset the running tickers. Folder lists, webserver listen_addr, urlbase, TLS, metrics, pprof, log settings, debug, quiet, parallel, and file modes return restartRequired: true; the daemon then re-execs itself once nothing is queued, extracting, or awaiting delete. In-flight extracts are never cancelled. Changed in v1.0.0 (September 2026).",
        "requestBody": {
          "required": true,
          "content": {"application/json": {"schema": {"$ref": "#/components/schemas/ConfigSectionPut"}}}
        },
        "responses": {
          "200": {
            "description": "Saved",
            "content": {"application/json": {"schema": {"$ref": "#/components/schemas/ConfigWriteReply"}}}
          },
          "400": {"$ref": "#/components/responses/BadRequest"},
          "500": {"$ref": "#/components/responses/PersistError"},
          "504": {"$ref": "#/components/responses/GatewayTimeout"},
          "404": {"description": "Unknown section", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/Error"}}}},
          "401": {"$ref": "#/components/responses/Unauthorized"},
          "403": {"$ref": "#/components/responses/Forbidden"}
        }
      }
    },
...
    "/api/config/{section}/test": {
      "parameters": [
        {
          "name": "section",
          "in": "path",
          "required": true,
          "schema": {
            "type": "string",
            "enum": ["general", "webserver", "sonarr", "radarr", "lidarr", "readarr", "folders", "hooks", "webhooks", "cmdhooks"]
          }
        }
      ],
      "post": {
        "tags": ["config"],
        "summary": "Probe a Starr instance or fire one sample hook",
        "description": "Requires config:{section}:write (same as PUT). Does not persist. Sonarr/Radarr/Lidarr/Readarr GET the activity queue on a throwaway client and return torrent vs usenet counts; blank url/apiKey are filled from the live instance matching slug. Webhooks and cmdhooks overlay non-empty posted fields onto a clone of the live hook (or a new hook); omitted shell/ignoreSsl keep live values. Fire one sample payload. general, webserver, folders, and hooks return 400. Remote Starr/hook failures are 424.",
        "requestBody": {
          "required": true,
          "content": {"application/json": {"schema": {"$ref": "#/components/schemas/ConfigTestRequest"}}}
        },
        "responses": {
          "200": {
            "description": "Starr queue counts or hook delivery result",
            "content": {"application/json": {"schema": {
              "anyOf": [
                {"$ref": "#/components/schemas/StarrTestResult"},
                {"$ref": "#/components/schemas/HookTestResult"}
              ]
            }}}
          },
          "400": {"$ref": "#/components/responses/BadRequest"},
          "404": {"description": "Unknown section", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/Error"}}}},
          "424": {"$ref": "#/components/responses/FailedDependency"},
          "401": {"$ref": "#/components/responses/Unauthorized"},
          "403": {"$ref": "#/components/responses/Forbidden"}
        }
      }
    }
  }
}